Bit of cleanup to dedupe a function.
