+++ apps/shoroukelkobrsi.com/src/components/homepage/get-cached-homepage.tsx
@@ -3,10 +3,8 @@ import { unstable_cache as cache } from 'next/cache'
 import configPromise from '@payload-config'
 import type { Homepage } from '@/payload-types'
 import type { Payload } from 'payload'
-import { createImageUrl } from '@/utilities/media'
-import styles from './reel.module.css'
 
-const getCachedHomepage = cache(
+export const getCachedHomepage = cache(
   async (): Promise<Homepage> => {
     const payload: Payload = await getPayloadHMR({
       config: configPromise,
@@ -23,16 +21,3 @@ const getCachedHomepage = cache(
     tags: ['homepage'],
   },
 )
-
-export default async function Reel(): Promise<React.ReactElement> {
-  const homepage = await getCachedHomepage()
-  const reel = homepage.reel
-
-  return (
-    <div className={styles.reelContainer}>
-      <video className={styles.reel} autoPlay muted loop>
-        <source src={createImageUrl(reel)} type="video/mp4" />
-      </video>
-    </div>
-  )
-}
